Document the fixed PATH requirement and share the canned sandbox test fake
The helper's module docstring now states that sh, stat, id, mkdir, chmod and any wrapped command must live in /usr/sbin:/usr/bin:/sbin:/bin, and what happens on an image that keeps them elsewhere.

FakeSandbox (injection tests) and ScriptedSandbox (helper tests) were near-identical canned-exec sandboxes; both now use CannedSandbox from tests/test_helpers/sandbox.py.
